pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "kadima"
version = "0.9.0"
description = "Hebrew NLP Corpus Pipeline — term extraction, validation, KB"
requires-python = ">=3.10"
license = {text = "MIT"}

# ── Core dependencies ────────────────────────────────────────────────────────
dependencies = [
    # Config / validation
    "PyYAML>=6.0,<7",
    "pydantic>=2.6,<3",

    # NLP (HebPipe — may require separate install)
    # "hebpipe>=0.2",  # uncomment when available on PyPI

    # HTTP clients (annotation + LLM backends)
    "httpx>=0.27,<1",

    # API server
    "fastapi>=0.110,<1",
    "uvicorn[standard]>=0.29,<1",
]

# ── Optional dependencies ────────────────────────────────────────────────────
[project.optional-dependencies]
gui = [
    "PyQt6>=6.6,<7",
]
dev = [
    "pytest>=8.0,<9",
    "pytest-cov>=5,<6",
    "ruff>=0.3,<1",
    "mypy>=1.8,<2",
]
all = [
    "kadima[gui,dev]",
]

[project.scripts]
kadima = "kadima.cli:main"

# ── Tool config ──────────────────────────────────────────────────────────────
[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]
ignore = ["E501"]  # line length handled by formatter

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
